# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.CreateQueueInput:
    properties:
      branchId:
        type: string
      customerName:
        type: string
      serviceId:
        type: string
    type: object
...
  handlers.UpdateQueueStatusInput:
    properties:
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: tangtnp@example.com
    name: Tang Tanathorn
  description: |-
    QueuePlus backend API for queue management system.
    This API provides authentication, branch management, service management, and queue operations.
  license:
    name: MIT
  title: QueuePlus API
  version: "1.0"
paths:
...
  /dashboard/queue-stats:
    get:
      description: Get queue statistics using MongoDB aggregation pipeline
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Queue dashboard stats
      tags:
      - Dashboard
  /external/posts:
    get:
      description: Fetch posts from an external API and transform the response
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties: true
            type: object
      summary: External posts
      tags:
      - External
...
  /queues:
    get:
      description: Get queues with filtering, pagination, and sorting
      parameters:
      - description: Branch ID
        in: query
        name: branchId
        type: string
      - description: Service ID
        in: query
        name: serviceId
        type: string
      - description: Queue status
        in: query
        name: status
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sortBy
        type: string
      - description: asc or desc
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: List queues
      tags:
      - Queues
    post:
      consumes:
      - application/json
      description: Create a new queue for a customer under a branch and service
      parameters:
      - description: Queue payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateQueueInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Create queue
      tags:
      - Queues
  /queues/{id}:
    get:
      description: Get a single queue by ID
      parameters:
      - description: Queue ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get queue by ID
      tags:
      - Queues
  /queues/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update queue status such as waiting, serving, done, or cancelled
      parameters:
      - description: Queue ID
        in: path
        name: id
        required: true
        type: string
      - description: Queue status payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateQueueStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update queue status
      tags:
      - Queues
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"